refactor: move test from fireEvent tests to render tests
