+++ src/__tests__/render.test.tsx
@@ -297,3 +297,29 @@ test('should throw an error when rerendering with text outside of Text component
     'Text strings must be rendered within a host Text component.'
   );
 });
+
+const ErrorComponent = () => {
+  const [shouldDisplayText, setShouldDisplayText] = React.useState(false);
+
+  if (!shouldDisplayText) {
+    return (
+      <TouchableOpacity
+        onPress={() => {
+          setShouldDisplayText(true);
+        }}
+      >
+        <Text>Display text</Text>
+      </TouchableOpacity>
+    );
+  }
+
+  return <View>text rendered outside text component</View>;
+};
+
+test('should throw an error when strings are rendered outside Text', () => {
+  const { getByText } = render(<ErrorComponent />);
+
+  expect(() => fireEvent.press(getByText('Display text'))).toThrowError(
+    'Text strings must be rendered within a host Text component.'
+  );
+});
